De-duplicate single axis
